# chore refs(DPLAN-12660): massively improve performance in docx export