--- a/demosplan/DemosPlanCoreBundle/Logic/Statement/StatementService.php
+++ b/demosplan/DemosPlanCoreBundle/Logic/Statement/StatementService.php
@@ -115,6 +115,7 @@ use demosplan\DemosPlanCoreBundle\ValueObject\AssessmentTable\StatementBulkEditV
 use demosplan\DemosPlanCoreBundle\ValueObject\ElasticsearchResultSet;
 use demosplan\DemosPlanCoreBundle\ValueObject\MovedStatementData;
 use demosplan\DemosPlanCoreBundle\ValueObject\PercentageDistribution;
+use demosplan\DemosPlanCoreBundle\ValueObject\Statement\StatementViewed;
 use demosplan\DemosPlanCoreBundle\ValueObject\StatementMovement;
 use demosplan\DemosPlanCoreBundle\ValueObject\StatementMovementCollection;
 use demosplan\DemosPlanCoreBundle\ValueObject\ToBy;
@@ -873,16 +874,18 @@ class StatementService extends CoreService implements StatementServiceInterface
     public function logStatementViewed($procedureId, $statements): void
     {
         try {
+            $entries = [];
             $accessMap = $this->generateAccessMap();
             if (0 !== count($accessMap) && 0 < sizeof($statements)) {
                 foreach ($statements as $statement) {
                     $publicStatement = $statement instanceof Statement ? $statement->getPublicStatement() : $statement['publicStatement'];
                     $statementId = $statement instanceof Statement ? $statement->getId() : $statement['id'];
                     if (0 < count($accessMap) && 0 === \strcmp((string) $publicStatement, (string) Statement::EXTERNAL)) {
-                        $this->addStatementViewedReport($procedureId, $accessMap, $statementId);
+                        $entries[] = new StatementViewed($procedureId, $accessMap, $statementId);
                     }
                 }
             }
+            $this->addStatementViewedReport($entries);
         } catch (Exception $e) {
             $this->logger->warning('protocol not saved: ', [$e]);
         }
@@ -1521,36 +1524,41 @@ class StatementService extends CoreService implements StatementServiceInterface
 
     /**
      * Add a report entry to the DB.
-     *
-     * @param string $procedureId
-     * @param array  $accessMap
-     * @param string $statementId
+     * @var StatementViewed[] $viewsToLog
      *
      * @throws ORMException
      * @throws OptimisticLockException
      * @throws ReflectionException
      */
-    private function addStatementViewedReport($procedureId, $accessMap, $statementId): void
+    private function addStatementViewedReport(array $viewsToLog): void
     {
-        // only log if user is known
-        if (!\array_key_exists('user', $accessMap)) {
-            return;
+        $entries = [];
+        foreach ($viewsToLog as $viewedEntry) {
+            $procedureId = $viewedEntry->getProcedureId();
+            $accessMap = $viewedEntry->getAccessMap();
+            $statementId = $viewedEntry->getStatementId();
+
+            // only log if user is known
+            if (!\array_key_exists('user', $accessMap)) {
+                return;
+            }
+            $alreadyLogged = $this->reportService
+                ->statementViewLogged($procedureId, $accessMap['user'], $statementId);
+
+            // logging access once is enough
+            if ($alreadyLogged) {
+                return;
+            }
+
+            $entries[] = $this->statementReportEntryFactory->createViewedEntry(
+                $statementId,
+                $procedureId,
+                $accessMap
+            );
+
         }
-        $alreadyLogged = $this->reportService
-            ->statementViewLogged($procedureId, $accessMap['user'], $statementId);
 
-        // logging access once is enough
-        if ($alreadyLogged) {
-            return;
-        }
-
-        $entry = $this->statementReportEntryFactory->createViewedEntry(
-            $statementId,
-            $procedureId,
-            $accessMap
-        );
-
-        $this->reportService->persistAndFlushReportEntries($entry);
+        $this->reportService->persistAndFlushReportEntries($entries);
     }
 
     /**
@@ -1629,7 +1637,7 @@ class StatementService extends CoreService implements StatementServiceInterface
         try {
             if (0 < count($accessMap) && 0 === \strcmp($statement->getPublicStatement(), (string) Statement::EXTERNAL)) {
                 try {
-                    $this->addStatementViewedReport($statement->getPId(), $accessMap, $statement->getId());
+                    $this->addStatementViewedReport([new StatementViewed($statement->getPId(), $accessMap, $statement->getId())]);
                 } catch (Exception $e) {
                     $this->logger->warning('Add Report in getStatementByIdent() failed Message: ', [$e]);
                 }
@@ -1664,7 +1672,7 @@ class StatementService extends CoreService implements StatementServiceInterface
                 $accessMap = $this->generateAccessMap();
                 if (0 < count($accessMap) && 0 === \strcmp($statement->getPublicStatement(), (string) Statement::EXTERNAL)) {
                     try {
-                        $this->addStatementViewedReport($statement->getPId(), $accessMap, $statement->getId());
+                        $this->addStatementViewedReport([new StatementViewed($statement->getPId(), $accessMap, $statement->getId())]);
                     } catch (Exception $e) {
                         $this->logger->warning('Add Report in getStatement() failed Message: ', [$e]);
                     }