### Ticket
DPLAN-12660

This change improves export performance by 82% by creating bulk report entries instead of single entries

![image](https://github.com/user-attachments/assets/ed1c3f4e-c52f-452f-a2bf-285a25b71699)
20.9 s → 3.79 s

for the complete export even more
![image](https://github.com/user-attachments/assets/c7b2cf5e-9ba6-4d90-a226-9ab40e6f7338)
12 min 19 s → 23.2 s

### How to review/test
Export statements in ATabelle

<!-- 
### Linked PRs (optional)
List other PRs that are somehow connected to this and explain the connection. Don't
link private repositories in public ones, but the other way around is fine.

- Other PR1 #{PR-number1}
- Other PR2 #{PR-number2}
-->

<!--
### Tasks (optional)
A list of all related tasks that need to be done before this can be merged.

- [x] Task1
- [ ] Task2
-->

### PR Checklist
<!-- Reminders for handling PRs -->

Delete the checkbox if it doesn't apply/isn't necessary.

- [ ] Tests updated/created
- [ ] Update documentation
- [ ] Link all relevant tickets
- [ ] Move the tickets on the board accordingly
- [ ] Data-Cy attributes added/updated ([conventions](https://dplan-documentation.demos-europe.eu/development/guidelines-conventions/coding-styleguides/twig_html.html#guideline-for-naming-cypress-hooks))